refactor(customer360): split god-component Customer360 (710→109)
Continua o split (format/hooks/components já extraídos): move o render inline
para 4 componentes de seção sob src/components/customer360/. O pai vira só
hooks de dados + revenueDerived + guards + composição.

- viewTypes.ts: tipos de view derivados dos hooks (Customer/CustomerScore/etc).
- CustomerHero.tsx: breadcrumb + hero (avatar/badges/status chips/ações); calcula
  health/churn internamente.
- CustomerKpiStrip.tsx: 4 KPIs; calcula fatTrend90 internamente.
- IdentityColumn.tsx: Contato + Contatos extras + Endereço + Score comercial.
- ActivityColumn.tsx: Itens preferidos + Últimos contatos + Pedidos recentes.
- +7 testes (4 arquivos): hero (PJ/PF/voltar), KPIs, identidade (contatos/endereço/
  score) e atividade (empty states/pedidos).

tsc --noEmit limpo; lint 0 problemas.
